+++ cmd/qsctl/mv.go
@@ -63,17 +63,21 @@ func mvRun(c *cobra.Command, args []string) (err error) {
 	}
 
 	go func() {
-		taskutils.StartProgress(time.Second, 3)
+		taskutils.StartProgress(time.Second)
 	}()
 	defer taskutils.FinishProgress()
 
 	if mvInput.Recursive {
 		t := task.NewMoveDir(rootTask)
+		t.SetHandleObjCallback(func(o *types.Object) {
+			fmt.Println(i18n.Sprintf("<%s> moved", o.Name))
+		})
 		t.Run()
 
 		if t.GetFault().HasError() {
 			return t.GetFault()
 		}
+
 		taskutils.WaitProgress()
 		i18n.Printf("Dir <%s> moved to <%s>.\n",
 			filepath.Join(srcWorkDir, t.GetSourcePath()), filepath.Join(dstWorkDir, t.GetDestinationPath()))
@@ -85,6 +89,7 @@ func mvRun(c *cobra.Command, args []string) (err error) {
 	if t.GetFault().HasError() {
 		return t.GetFault()
 	}
+
 	taskutils.WaitProgress()
 	i18n.Printf("File <%s> moved to <%s>.\n",
 		filepath.Join(srcWorkDir, t.GetSourcePath()), filepath.Join(dstWorkDir, t.GetDestinationPath()))